anomaly-detector/requirements.txt
kafka-python===1.4.5
nupic===1.0.5
